+++ test/tracing/reactnativenavigation.test.ts
@@ -1,22 +1,25 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import type { TransactionContext } from '@sentry/types';
+import type { Transaction, TransactionContext } from '@sentry/types';
 import type { EmitterSubscription } from 'react-native';
 
 import type {
+  BottomTabPressedEvent,
   ComponentWillAppearEvent,
   EventsRegistry,
   NavigationDelegate} from '../../src/js/tracing/reactnativenavigation';
 import {
   ReactNativeNavigationInstrumentation,
 } from '../../src/js/tracing/reactnativenavigation';
-import { getBlankTransactionContext } from '../../src/js/tracing/utils';
+import type { TransactionCreator } from '../../src/js/tracing/routingInstrumentation';
 import { getMockTransaction } from '../testutils';
 
 interface MockEventsRegistry extends EventsRegistry {
   componentWillAppearListener?: (event: ComponentWillAppearEvent) => void;
   commandListener?: (name: string, params: unknown) => void;
+  bottomTabPressedListener?: (event: BottomTabPressedEvent) => void;
   onComponentWillAppear(event: ComponentWillAppearEvent): void;
   onCommand(name: string, params: unknown): void;
+  onBottomTabPressed(event: BottomTabPressedEvent): void;
 }
 
 const mockEventsRegistry: MockEventsRegistry = {
@@ -25,6 +28,9 @@ const mockEventsRegistry: MockEventsRegistry = {
   },
   onCommand(name: string, params: unknown): void {
     this.commandListener?.(name, params);
+  },
+  onBottomTabPressed(event) {
+    this.bottomTabPressedListener?.(event);
   },
   registerComponentWillAppearListener(
     callback: (event: ComponentWillAppearEvent) => void
@@ -42,6 +48,13 @@ const mockEventsRegistry: MockEventsRegistry = {
       remove() {},
     };
   },
+  registerBottomTabPressedListener(callback: (event: BottomTabPressedEvent) => void) {
+    this.bottomTabPressedListener = callback;
+    return {
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      remove() { },
+    } as EmitterSubscription;
+  },
 };
 
 const mockNavigationDelegate: NavigationDelegate = {
@@ -51,21 +64,32 @@ const mockNavigationDelegate: NavigationDelegate = {
 };
 
 describe('React Native Navigation Instrumentation', () => {
-  test('Correctly instruments a route change', () => {
-    const instrumentation = new ReactNativeNavigationInstrumentation(
+  let instrumentation: ReactNativeNavigationInstrumentation;
+  let tracingListener: jest.Mock<ReturnType<TransactionCreator>, Parameters<TransactionCreator>>;
+  let mockTransaction: Transaction;
+
+  beforeEach(() => {
+    instrumentation = new ReactNativeNavigationInstrumentation(
       mockNavigationDelegate
     );
 
-    const mockTransaction = getMockTransaction(
+    mockTransaction = getMockTransaction(
       ReactNativeNavigationInstrumentation.instrumentationName
     );
-    const tracingListener = jest.fn(() => mockTransaction);
+
+    tracingListener = jest.fn((_context: TransactionContext) => mockTransaction);
     instrumentation.registerRoutingInstrumentation(
       tracingListener,
       (context) => context,
       () => {}
     );
+  });
 
+  afterEach(() => {
+    jest.resetAllMocks();
+  });
+
+  test('Correctly instruments a route change', () => {
     mockEventsRegistry.onCommand('root', {});
 
     expect(mockTransaction.name).toBe('Route Change');
@@ -78,33 +102,30 @@ describe('React Native Navigation Instrumentation', () => {
     };
     mockEventsRegistry.onComponentWillAppear(mockEvent);
 
-    expect(mockTransaction.name).toBe(mockEvent.componentName);
-    expect(mockTransaction.tags).toStrictEqual({
-      ...getBlankTransactionContext(
-        ReactNativeNavigationInstrumentation.instrumentationName
-      ).tags,
-      'routing.route.name': mockEvent.componentName,
-    });
-    expect(mockTransaction.data).toStrictEqual({
-      route: {
-        ...mockEvent,
-        name: mockEvent.componentName,
-        hasBeenSeen: false,
+    expect(mockTransaction).toEqual(expect.objectContaining(<Partial<Transaction>>{
+      name: 'Test',
+      tags: {
+        'routing.instrumentation': 'react-native-navigation',
+        'routing.route.name': 'Test',
       },
-      previousRoute: null,
-    });
-    expect(mockTransaction.metadata.source).toBe('component');
+      data: {
+        route: {
+          componentId: '0',
+          componentName: 'Test',
+          componentType: 'Component',
+          passProps: {},
+          name: 'Test',
+          hasBeenSeen: false,
+        },
+        previousRoute: null,
+      },
+      metadata: expect.objectContaining({
+        source: 'component',
+      }),
+    }));
   });
 
   test('Transaction context is changed with beforeNavigate', () => {
-    const instrumentation = new ReactNativeNavigationInstrumentation(
-      mockNavigationDelegate
-    );
-
-    const mockTransaction = getMockTransaction(
-      ReactNativeNavigationInstrumentation.instrumentationName
-    );
-    const tracingListener = jest.fn(() => mockTransaction);
     instrumentation.registerRoutingInstrumentation(
       tracingListener,
       (context) => {
@@ -129,42 +150,33 @@ describe('React Native Navigation Instrumentation', () => {
     };
     mockEventsRegistry.onComponentWillAppear(mockEvent);
 
-    expect(mockTransaction.name).toBe('New Name');
-    expect(mockTransaction.description).toBe('Description');
-    expect(mockTransaction.sampled).toBe(false);
-    expect(mockTransaction.tags).toStrictEqual({
-      ...getBlankTransactionContext(
-        ReactNativeNavigationInstrumentation.instrumentationName
-      ).tags,
-      'routing.route.name': mockEvent.componentName,
-    });
-    expect(mockTransaction.data).toStrictEqual({
-      route: {
-        ...mockEvent,
-        name: mockEvent.componentName,
-        hasBeenSeen: false,
+    expect(mockTransaction).toEqual(expect.objectContaining(<Partial<Transaction>>{
+      name: 'New Name',
+      description: 'Description',
+      sampled: false,
+      tags: {
+        'routing.instrumentation': 'react-native-navigation',
+        'routing.route.name': 'Test',
       },
-      previousRoute: null,
-    });
-    expect(mockTransaction.metadata.source).toBe('custom');
+      data: {
+        route: {
+          componentId: '0',
+          componentName: 'Test',
+          componentType: 'Component',
+          passProps: {},
+          name: 'Test',
+          hasBeenSeen: false,
+        },
+        previousRoute: null,
+      },
+      metadata: expect.objectContaining({
+        source: 'custom',
+      }),
+    }));
   });
 
   test('Transaction not sent on a cancelled route change', () => {
     jest.useFakeTimers();
-
-    const instrumentation = new ReactNativeNavigationInstrumentation(
-      mockNavigationDelegate
-    );
-
-    const mockTransaction = getMockTransaction(
-      ReactNativeNavigationInstrumentation.instrumentationName
-    );
-    const tracingListener = jest.fn(() => mockTransaction);
-    instrumentation.registerRoutingInstrumentation(
-      tracingListener,
-      (context) => context,
-      () => {}
-    );
 
     mockEventsRegistry.onCommand('root', {});
 
@@ -180,21 +192,6 @@ describe('React Native Navigation Instrumentation', () => {
 
   test('Transaction not sent if route change timeout is passed', () => {
     jest.useFakeTimers();
-
-    const instrumentation = new ReactNativeNavigationInstrumentation(
-      mockNavigationDelegate,
-      { routeChangeTimeoutMs: 500 }
-    );
-
-    const mockTransaction = getMockTransaction(
-      ReactNativeNavigationInstrumentation.instrumentationName
-    );
-    const tracingListener = jest.fn(() => mockTransaction);
-    instrumentation.registerRoutingInstrumentation(
-      tracingListener,
-      (context) => context,
-      () => {}
-    );
 
     mockEventsRegistry.onCommand('root', {});
 
@@ -217,17 +214,79 @@ describe('React Native Navigation Instrumentation', () => {
     jest.useRealTimers();
   });
 
-  describe('onRouteConfirmed', () => {
-    test('onRouteConfirmed called with correct route data', () => {
-      const instrumentation = new ReactNativeNavigationInstrumentation(
-        mockNavigationDelegate
+  describe('tab change', () => {
+    beforeEach(() => {
+      instrumentation = new ReactNativeNavigationInstrumentation(mockNavigationDelegate, { enableTabsInstrumentation: true });
+      instrumentation.registerRoutingInstrumentation(
+        tracingListener,
+        (context) => context,
+        () => { },
+      );
+    });
+
+    test('correctly instruments a tab change', () => {
+      mockEventsRegistry.onBottomTabPressed({ tabIndex: 0 });
+      mockEventsRegistry.onComponentWillAppear(<ComponentWillAppearEvent>{
+        componentId: '0',
+        componentName: 'TestScreenName',
+        componentType: 'Component',
+        passProps: {},
+      });
+
+      expect(mockTransaction.toContext()).toEqual(expect.objectContaining(<Partial<TransactionContext>>{
+        name: 'TestScreenName',
+        tags: {
+          'routing.instrumentation': 'react-native-navigation',
+          'routing.route.name': 'TestScreenName',
+        },
+        data: {
+          route: {
+            componentId: '0',
+            componentName: 'TestScreenName',
+            componentType: 'Component',
+            passProps: {},
+            name: 'TestScreenName',
+            hasBeenSeen: false,
+          },
+          previousRoute: null,
+        },
+      }));
+    });
+
+    test('not instrument tabs if disabled', () => {
+      jest.useFakeTimers();
+
+      instrumentation = new ReactNativeNavigationInstrumentation(
+        mockNavigationDelegate,
+        { enableTabsInstrumentation: false },
+      );
+      tracingListener = jest.fn((_context: TransactionContext) => mockTransaction);
+      instrumentation.registerRoutingInstrumentation(
+        tracingListener,
+        (context) => context,
+        () => { },
       );
 
-      const mockTransaction = getMockTransaction(
-        ReactNativeNavigationInstrumentation.instrumentationName
-      );
-      const tracingListener = jest.fn(() => mockTransaction);
-      let confirmedContext: TransactionContext | undefined;
+      mockEventsRegistry.onBottomTabPressed({ tabIndex: 0 });
+      mockEventsRegistry.onComponentWillAppear(<ComponentWillAppearEvent>{
+        componentId: '0',
+        componentName: 'TestScreenName',
+        componentType: 'Component',
+      });
+
+      expect(tracingListener).not.toBeCalled();
+
+      jest.runAllTimers();
+      expect(mockTransaction.sampled).toBe(false);
+
+      jest.useRealTimers();
+    });
+  });
+
+  describe('onRouteConfirmed', () => {
+    let confirmedContext: TransactionContext | undefined;
+
+    beforeEach(() => {
       instrumentation.registerRoutingInstrumentation(
         tracingListener,
         (context) => context,
@@ -235,7 +294,9 @@ describe('React Native Navigation Instrumentation', () => {
           confirmedContext = context;
         }
       );
+    });
 
+    test('onRouteConfirmed called with correct route data', () => {
       mockEventsRegistry.onCommand('root', {});
 
       expect(mockTransaction.name).toBe('Route Change');
@@ -258,43 +319,20 @@ describe('React Native Navigation Instrumentation', () => {
       };
       mockEventsRegistry.onComponentWillAppear(mockEvent2);
 
-      expect(confirmedContext).toBeDefined();
-      if (confirmedContext) {
-        expect(confirmedContext.name).toBe(mockEvent2.componentName);
-        expect(confirmedContext.metadata).toBeUndefined();
-        expect(confirmedContext.data).toBeDefined();
-        if (confirmedContext.data) {
-          expect(confirmedContext.data.route.name).toBe(
-            mockEvent2.componentName
-          );
-          expect(confirmedContext.data.previousRoute).toBeDefined();
-          if (confirmedContext.data.previousRoute) {
-            expect(confirmedContext.data.previousRoute.name).toBe(
-              mockEvent1.componentName
-            );
-          }
-        }
-      }
+      expect(confirmedContext).toEqual(expect.objectContaining(<Partial<TransactionContext>>{
+        name: 'Test 2',
+        data: {
+          route: expect.objectContaining({
+            name: 'Test 2',
+          }),
+          previousRoute: expect.objectContaining({
+            name: 'Test 1',
+          }),
+        },
+      }));
     });
 
     test('onRouteConfirmed clears transaction', () => {
-      const instrumentation = new ReactNativeNavigationInstrumentation(
-        mockNavigationDelegate
-      );
-
-      const mockTransaction = getMockTransaction(
-        ReactNativeNavigationInstrumentation.instrumentationName
-      );
-      const tracingListener = jest.fn(() => mockTransaction);
-      let confirmedContext: TransactionContext | undefined;
-      instrumentation.registerRoutingInstrumentation(
-        tracingListener,
-        (context) => context,
-        (context) => {
-          confirmedContext = context;
-        }
-      );
-
       mockEventsRegistry.onCommand('root', {});
 
       expect(mockTransaction.name).toBe('Route Change');
@@ -315,10 +353,9 @@ describe('React Native Navigation Instrumentation', () => {
       };
       mockEventsRegistry.onComponentWillAppear(mockEvent2);
 
-      expect(confirmedContext).toBeDefined();
-      if (confirmedContext) {
-        expect(confirmedContext.name).toBe(mockEvent1.componentName);
-      }
+      expect(confirmedContext).toEqual(expect.objectContaining(<Partial<TransactionContext>>{
+        name: 'Test 1',
+      }));
     });
   });
 });